Modelify Lines; very early implementation for reading status
Each line is now a model and has a list of readers. Adding to or
removing from this list automatically redraws that line in the UI.
Reading notifications are handled by the Room, which distributes those
to the lines.

There are no tests yet for this functionality, also I have not run lint
on it yet.

This currently fakes a lot of stuff in js which should ideally live in
the server but is not currently implemented.

This also includes first styling attempts using stylus, thats still open
for discussion though